Separated out actor code, started on test parallelism
