+++ data/todos.json
@@ -1,5 +1,10 @@
 [
   {
+    "id": 33,
+    "description": "new record via bruno",
+    "status": "NOT STARTED"
+  },
+  {
     "id": 32,
     "description": "this is body number 9",
     "status": "NOT STARTED"
@@ -35,11 +40,6 @@
     "status": "NOT STARTED"
   },
   {
-    "id": 24,
-    "description": "this is body number 1",
-    "status": "NOT STARTED"
-  },
-  {
     "id": 23,
     "description": "this is body number 0",
     "status": "NOT STARTED"
@@ -71,8 +71,8 @@
   },
   {
     "id": 17,
-    "description": "another new record for monday a.m",
-    "status": "NOT STARTED"
+    "description": "update for 17 - via bruno again",
+    "status": "STARTED"
   },
   {
     "id": 15,
@@ -90,8 +90,8 @@
     "status": "NOT STARTED"
   },
   {
-    "id": 33,
-    "description": "update for 33",
-    "status": "STARTED"
+    "id": 34,
+    "description": "new record via bruno on sunday",
+    "status": "NOT STARTED"
   }
 ]
